requirements.txt
flask==2.3.2
tensorflow==2.13.0
numpy==1.23.5
Pillow==9.5.0